allow sent multiple contact request / hints on PairWithOthersScreen / replace ContactPairingStatus by IntDef
