+++ app/src/main/java/tech/relaycorp/letro/pairing/ui/PairWithOthersScreen.kt
@@ -24,6 +24,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import tech.relaycorp.letro.R
 import tech.relaycorp.letro.pairing.PairWithOthersViewModel
 import tech.relaycorp.letro.ui.common.LetroButtonMaxWidthFilled
+import tech.relaycorp.letro.ui.common.LetroInfoView
 import tech.relaycorp.letro.ui.common.LetroOutlinedTextField
 import tech.relaycorp.letro.ui.theme.HorizontalScreenPadding
 import tech.relaycorp.letro.ui.theme.LetroTheme
@@ -39,6 +40,8 @@ fun PairWithOthersScreen(
             onBackClick()
         }
     }
+
+    val errorCaption = uiState.pairingErrorCaption
 
     Column(
         modifier = Modifier.padding(
@@ -76,7 +79,29 @@ fun PairWithOthersScreen(
             onValueChange = viewModel::onIdChanged,
             label = R.string.general_id,
             hintText = stringResource(id = R.string.new_contact_id_hint),
-        )
+            isError = errorCaption != null,
+        ) {
+            if (errorCaption != null) {
+                Spacer(modifier = Modifier.height(6.dp))
+                Text(
+                    text = stringResource(id = errorCaption.message),
+                    color = MaterialTheme.colorScheme.error,
+                    style = MaterialTheme.typography.bodySmall,
+                )
+            }
+        }
+        if (uiState.isSentRequestAgainHintVisible) {
+            Spacer(
+                modifier = Modifier.height(8.dp),
+            )
+            LetroInfoView {
+                Text(
+                    text = stringResource(id = R.string.pair_request_was_already_sent_hint),
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurface,
+                )
+            }
+        }
         Spacer(
             modifier = Modifier.height(24.dp),
         )
@@ -92,6 +117,7 @@ fun PairWithOthersScreen(
         LetroButtonMaxWidthFilled(
             text = stringResource(id = R.string.onboarding_pair_with_people_button),
             onClick = { viewModel.onPairRequestClick() },
+            isEnabled = errorCaption == null,
         )
     }
 }
