+++ app/src/main/java/tech/relaycorp/letro/storage/LetroDatabase.kt
@@ -2,12 +2,10 @@ package tech.relaycorp.letro.storage
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import androidx.room.TypeConverters
 import tech.relaycorp.letro.account.model.Account
 import tech.relaycorp.letro.account.storage.AccountDao
 import tech.relaycorp.letro.contacts.model.Contact
 import tech.relaycorp.letro.contacts.storage.ContactsDao
-import tech.relaycorp.letro.contacts.storage.converter.ContactPairingStatusConverter
 
 @Database(
     entities = [
@@ -16,7 +14,6 @@ import tech.relaycorp.letro.contacts.storage.converter.ContactPairingStatusConve
     ],
     version = 1,
 )
-@TypeConverters(ContactPairingStatusConverter::class)
 abstract class LetroDatabase : RoomDatabase() {
     abstract fun accountDao(): AccountDao
     abstract fun contactsDao(): ContactsDao
