+++ app/src/main/java/tech/relaycorp/letro/ui/common/LetroInfoView.kt
@@ -1,0 +1,43 @@
+package tech.relaycorp.letro.ui.common
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import tech.relaycorp.letro.R
+
+@Composable
+fun LetroInfoView(
+    content: @Composable () -> Unit,
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(
+                color = MaterialTheme.colorScheme.surfaceVariant,
+                shape = RoundedCornerShape(8.dp),
+            )
+            .padding(8.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Icon(
+            painter = painterResource(id = R.drawable.info),
+            tint = MaterialTheme.colorScheme.onPrimaryContainer,
+            contentDescription = null,
+        )
+        Spacer(
+            modifier = Modifier.width(8.dp),
+        )
+        content()
+    }
+}
