+++ app/src/main/res/drawable-v24/arrow_back.xml
@@ -5,8 +5,10 @@
     android:viewportHeight="18">
   <path
       android:pathData="M0,9C0,8.379 0.504,7.875 1.125,7.875H16.875C17.496,7.875 18,8.379 18,9C18,9.621 17.496,10.125 16.875,10.125H1.125C0.504,10.125 0,9.621 0,9Z"
-      android:fillColor="#0C1B44" />
+      android:fillColor="#0C1B44"
+      android:fillType="evenOdd"/>
   <path
       android:pathData="M9.795,0.33C10.235,0.769 10.235,1.481 9.795,1.921L2.716,9L9.795,16.08C10.235,16.519 10.235,17.231 9.795,17.67C9.356,18.11 8.644,18.11 8.205,17.67L0.33,9.795C-0.11,9.356 -0.11,8.644 0.33,8.205L8.205,0.33C8.644,-0.11 9.356,-0.11 9.795,0.33Z"
-      android:fillColor="#0C1B44" />
+      android:fillColor="#0C1B44"
+      android:fillType="evenOdd"/>
 </vector>
