+++ app/src/main/res/values/strings.xml
@@ -26,10 +26,6 @@
     <string name="onboarding_pairing_request_sent_message">You’ll receive a notification once you and %s are connected.</string>
     <string name="onboarding_pairing_request_sent_button">Got it</string>
 
-    <string name="onboarding_use_existing_account_domain">Domain</string>
-    <string name="onboarding_use_existing_account_token">Token</string>
-    <string name="onboarding_use_existing_account_button">Confirm details</string>
-
     <string name="onboarding_pair_with_people_alias">Name (optional)</string>
     <string name="onboarding_pair_with_people_button">Request pairing</string>
 
@@ -41,6 +37,10 @@
 
     <string name="no_browser">There is no app to install Awala</string>
     <string name="no_app_to_share">There is no app to share your id</string>
+
+    <string name="pair_request_was_already_sent_hint">You already tried to connect with them, but we can send another request if you want.</string>
+    <string name="pair_request_already_paired">You two are already connected.</string>
+    <string name="pair_request_already_in_progress">Your pairing with this user is already under way.</string>
 
     <string name="join_me_on_letro">Join me on Letro: https://letro.app/connect/#u=%1$s</string>
